refactor(playtest): PlaytestUiOpsをUi配下3ファイルへ分割しwebui-buildmenu-clickを復旧する
A9再レビューのImportant B対応: 280行だったPlaytestUiOpsを責務ごとに分割し、
Operations直下が10ファイル上限のためOperations/Uiサブディレクトリを新設した。
- Ui/PlaytestUiOps.cs (147行): UI状態観測・照準・クリック/ドラッグ設置
- Ui/PlaytestBuildMenuOps.cs (100行): ビルドメニューを開いてブロックを選ぶ経路
- Ui/PlaytestSkitOps.cs (62行): 開幕スキットのskip前提

併せてImportant A-2の実測: webui-buildmenu-clickのredは「カテゴリ限定描画」ではなく
(1)ブロック未アンロックで一覧から除外 (2)全カテゴリが1本のスクロールに並ぶため
エントリがビューポート外(y=1793)でヒットテストに落ちる、の2点だった。
アンロックとカテゴリ見出しクリックを足して green (3/3) に復旧した。
