+++ .agents/skills/unity-playmode-recorded-playtest/scenarios/building/placement-reason-tooltip.cs
@@ -8,6 +8,7 @@ using Client.Game.InGame.UI.Tooltip;
 using Client.Playtest;
 using Client.Playtest.Input;
 using Client.Playtest.Operations;
+using Client.Playtest.Operations.Ui;
 using Cysharp.Threading.Tasks;
 using Game.Block.Interface;
 using System.Linq;
