+++ moorestech_client/Assets/Scripts/Client.Playtest/Operations/PlaytestHotbarOps.cs
@@ -4,6 +4,7 @@ using Client.Game.InGame.Context;
 using Client.Game.InGame.Hotbar;
 using Client.Game.InGame.UI.UIState;
 using Client.Playtest.Input;
+using Client.Playtest.Operations.Ui;
 using Core.Master;
 using Cysharp.Threading.Tasks;
 using Game.Context;
