+++ moorestech_client/Assets/Scripts/Client.Playtest/Operations/Ui.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48dcbab8e7c3449d2ba40a75c2da0f9a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
