+++ moorestech_client/Assets/Scripts/Client.Playtest/PlaytestDriver.cs
@@ -7,6 +7,7 @@ using Client.Game.InGame.UI.UIState;
 using Client.Playtest.Core;
 using Client.Playtest.Input;
 using Client.Playtest.Operations;
+using Client.Playtest.Operations.Ui;
 using Cysharp.Threading.Tasks;
 using Game.Block.Interface;
 using Game.Context;
@@ -129,13 +130,13 @@ namespace Client.Playtest
             await PlaytestUiOps.WaitUiState(state, timeoutSeconds);
             _reporter.EndWait(waitEntry);
         }
-        public async UniTask OpenBuildMenuAndSelectBlock(string blockName) => await _reporter.Act($"ビルドメニューで選択: {blockName}", () => PlaytestUiOps.OpenBuildMenuAndSelectBlock(blockName));
+        public async UniTask OpenBuildMenuAndSelectBlock(string blockName) => await _reporter.Act($"ビルドメニューで選択: {blockName}", () => PlaytestBuildMenuOps.OpenBuildMenuAndSelectBlock(blockName));
         // 開幕スキットを飛ばしゲーム画面まで抜ける。ホットバー・ビルドメニュー操作すべての前提
         // Skips the opening skit and lands on the game screen; the precondition for every hotbar and build-menu operation
-        public async UniTask SkipOpeningSkit() => await _reporter.Act("開幕スキットをSkipインテントで飛ばす", () => PlaytestUiOps.SkipOpeningSkit(SkitSkipTimeoutSeconds, SkitSkipUiStateTimeoutSeconds));
+        public async UniTask SkipOpeningSkit() => await _reporter.Act("開幕スキットをSkipインテントで飛ばす", () => PlaytestSkitOps.SkipOpeningSkit(SkitSkipTimeoutSeconds, SkitSkipUiStateTimeoutSeconds));
         // スキットが再生されないワールドでも落ちない版。再生されていれば飛ばし、されていなければそのまま進む
         // Variant that tolerates worlds without an opening skit: skips it when it plays, moves on when it does not
-        public async UniTask SkipOpeningSkitIfPlaying() => await _reporter.Act("開幕スキットを再生中なら飛ばす", () => PlaytestUiOps.SkipOpeningSkitIfPlaying(OptionalSkitSkipTimeoutSeconds));
+        public async UniTask SkipOpeningSkitIfPlaying() => await _reporter.Act("開幕スキットを再生中なら飛ばす", () => PlaytestSkitOps.SkipOpeningSkitIfPlaying(OptionalSkitSkipTimeoutSeconds));
         public async UniTask ExitToGameScreen() => await _reporter.Act("ゲーム画面へ戻る", PlaytestUiOps.ExitToGameScreen);
         public async UniTask AimAt(Vector3 worldPosition) => await _reporter.Act($"照準: {worldPosition}", () => PlaytestUiOps.AimAtWorldPosition(worldPosition));
         // 指定originに設置されるよう接地面上のフットプリント中心へ照準する（向きはNorth前提）
